passing the code to cpp libraries
Creating a 127 bytes max circular buffer library to gasp how to declare c++ libraries.
This lib may or may not be used in the project as a small FIFO or LIFO only if it proves more efficient than the built-in Stream class of arduino.
Next is to pass the code for the scanner from C lib to a C++ class library, and start work in the lexer after some clean up.